Move logic files to library
